productivity_tools_1/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.0
Flask-Migrate==4.0.5
bcrypt==4.1.2
email-validator==2.1.1
